refactor: Move wardrobe specific components into separate wardrobe folder
