+++ src/pages/wardrobe/components/interface.ts
